# Optimize distributed.Cache.Get when lookaside cache is enabled