--- a/enterprise/server/backends/distributed/distributed.go
+++ b/enterprise/server/backends/distributed/distributed.go
@@ -329,6 +329,7 @@ func (t *teeReadCloser) Read(p []byte) (n int, err error) {
 	}
 	return
 }
+
 func (t *teeReadCloser) Close() error {
 	err := t.rc.Close()
 	if err == nil {
@@ -336,6 +337,27 @@ func (t *teeReadCloser) Close() error {
 		_ = t.cwc.Close()
 	}
 	return err
+}
+
+type teeReadCloseWriterTo struct {
+	teeReadCloser
+	src io.WriterTo
+}
+
+func (t *teeReadCloseWriterTo) WriteTo(w io.Writer) (int64, error) {
+	return t.src.WriteTo(teeWriter{w, t.cwc})
+}
+
+type teeWriter struct {
+	a, b io.Writer
+}
+
+func (w teeWriter) Write(b []byte) (int, error) {
+	n, err := w.a.Write(b)
+	if err != nil {
+		return n, err
+	}
+	return w.b.Write(b[:n])
 }
 
 // lookasideKey returns the resource's key in the lookaside cache and true,
@@ -368,17 +390,21 @@ func (c *Cache) addLookasideEntry(r *rspb.ResourceName, data []byte) {
 		c.log.Debugf("Not setting lookaside entry for resource: %s", r)
 		return
 	}
+	c.setLookasideEntry(k, data)
+}
+
+func (c *Cache) setLookasideEntry(lookasideKey string, data []byte) {
 	entry := lookasideCacheEntry{
 		createdAtMillis: time.Now().UnixMilli(),
 		data:            data,
 	}
 
 	c.lookasideMu.Lock()
-	if !c.lookaside.Contains(k) {
-		c.lookaside.Add(k, entry)
+	if !c.lookaside.Contains(lookasideKey) {
+		c.lookaside.Add(lookasideKey, entry)
 	}
 	c.lookasideMu.Unlock()
-	c.log.Debugf("Set %q in lookaside cache", k)
+	c.log.Debugf("Set %q in lookaside cache", lookasideKey)
 }
 
 // getLookasideEntry returns the resource and if it was found in the lookaside
@@ -421,11 +447,11 @@ func (c *Cache) getLookasideEntry(r *rspb.ResourceName) ([]byte, bool) {
 	return nil, false
 }
 
-func (c *Cache) lookasideWriter(r *rspb.ResourceName) (interfaces.CommittedWriteCloser, error) {
-	buffer := new(bytes.Buffer)
+func (c *Cache) lookasideWriter(r *rspb.ResourceName, lookasideKey string) (interfaces.CommittedWriteCloser, error) {
+	buffer := bytes.NewBuffer(make([]byte, 0, bufferSize(r)))
 	wc := ioutil.NewCustomCommitWriteCloser(buffer)
 	wc.CommitFn = func(int64) error {
-		c.addLookasideEntry(r, buffer.Bytes())
+		c.setLookasideEntry(lookasideKey, buffer.Bytes())
 		return nil
 	}
 	return wc, nil
@@ -442,11 +468,19 @@ func (c *Cache) teeReadCloser(r *rspb.ResourceName, rc io.ReadCloser) io.ReadClo
 	if r.GetDigest().GetSizeBytes() > *maxLookasideEntryBytes {
 		return rc
 	}
-	lwc, err := c.lookasideWriter(r)
+	k, ok := lookasideKey(r)
+	if !ok {
+		return rc
+	}
+	lwc, err := c.lookasideWriter(r, k)
 	if err != nil {
 		return rc
 	}
-	return &teeReadCloser{rc, lwc}
+	res := teeReadCloser{rc, lwc}
+	if wt, ok := rc.(io.WriterTo); ok {
+		return &teeReadCloseWriterTo{res, wt}
+	}
+	return &res
 }
 
 func (c *Cache) recvHeartbeatCallback(ctx context.Context, peer string) {
@@ -1093,34 +1127,35 @@ func (c *Cache) distributedReader(ctx context.Context, rn *rspb.ResourceName, of
 	return nil, status.NotFoundErrorf("Exhausted all peers attempting to read %q.", rn.GetDigest().GetHash())
 }
 
-// Below, in Get(), this value is the max initial allocatable buffer size.
-// Set it somewhat conservatively so that we're not DOSed by someone crafting
-// remote_instance_names that match this just to use memory.
-const maxInitialByteBufferSize = (1024 * 1024 * 4)
+func bufferSize(rn *rspb.ResourceName) int {
+	if rn.GetCacheType() == rspb.CacheType_CAS {
+		// If this is a CAS object, size the buffer to fit exactly.
+		// Clamp the size between 0 and 10MB, to protect from invalid and
+		// malicious requests.
+		return min(10_000_000, max(0, int(rn.GetDigest().GetSizeBytes())))
+	} else if strings.HasPrefix(rn.GetInstanceName(), content_addressable_storage_server.TreeCacheRemoteInstanceName) {
+		// If this is a TreeCache entry that we wrote; pull the size
+		// from the remote instance name.
+		parts := strings.Split(rn.GetInstanceName(), "/")
+		if s, err := strconv.Atoi(parts[len(parts)-1]); err == nil {
+			// Limit this to 4MiB so that we're not DOSed by someone crafting
+			// remote_instance_names that match this just to use memory.
+			return min(1024*1024*4, s)
+		} else {
+			return 0
+		}
+	}
+	// The median and average AC results are less than 4KiB: go/action-result-size
+	return 4 * 1024
 
+}
 func (c *Cache) Get(ctx context.Context, rn *rspb.ResourceName) ([]byte, error) {
 	r, err := c.distributedReader(ctx, rn, 0, 0, "Get" /*=metricsLabel*/)
 	if err != nil {
 		return nil, err
 	}
 	defer r.Close()
-
-	var buf *bytes.Buffer
-	if rn.GetCacheType() == rspb.CacheType_CAS {
-		// If this is a CAS object, size the buffer to fit exactly.
-		buf = bytes.NewBuffer(make([]byte, 0, int(rn.GetDigest().GetSizeBytes())))
-	} else if strings.HasPrefix(rn.GetInstanceName(), content_addressable_storage_server.TreeCacheRemoteInstanceName) {
-		// If this is a TreeCache entry that we wrote; pull the size
-		// from the remote instance name.
-		parts := strings.Split(rn.GetInstanceName(), "/")
-		if s, err := strconv.Atoi(parts[len(parts)-1]); err == nil {
-			buf = bytes.NewBuffer(make([]byte, 0, min(s, maxInitialByteBufferSize)))
-		} else {
-			buf = new(bytes.Buffer)
-		}
-	} else {
-		buf = new(bytes.Buffer)
-	}
+	buf := bytes.NewBuffer(make([]byte, 0, bufferSize(rn)))
 	_, err = io.Copy(buf, r)
 	return buf.Bytes(), err
 }