--- a/enterprise/server/test/performance/cache/BUILD
+++ b/enterprise/server/test/performance/cache/BUILD
@@ -22,6 +22,7 @@ go_test(
         "//server/testutil/testport",
         "//server/util/log",
         "//server/util/prefix",
+        "//server/util/testing/flags",
     ],
 )