@vanja-p 2025-05-21
> Sorry for being slow here--i'm theoretically fine with this change, but there are some conflicts that you'll need to sync past now.

I handled the conflicts and simplified this change. Turns out that implementing WriterTo here doesn't help. Also, I reverted most of the buffer size changes, but I did presize the `lookasideWriter`, since it's a trivial change.